Add files via upload
Automated-Mailing-System
This is project is on email mailing system which is build using python and tkinter. The purpose of this project was to send mail to various employees reminding them about their holidays and weekend off’s. Information of various employees are stored in the excel sheet and when we run this program, the user interface pops up and sender can type his/her subject and body and then after clicking the send button, the employees whose record was in excel sheet will start receiving mail and after the completion of this task interface will display a alert message stating the email is sent to everyone.

Library’s Used
· pandas: to import the excel file storing the name and email address of the employee.

· smtplib: it is used to login to our email id and provide it access to send our mail through python code.

· Tkinter: It is used to create user interfaces and to create fields for email’s subject and body . A send button is also created to send emails at all email ids that were there in excel sheet.

To run the program
python final.py